move RecyclerViewLifecycleObserver to androidx-recyclerview module
